requirements.txt
bs4
python-dotenv
python-telegram-bot
pytz
requests
schedule
